voice-assistant-mvp/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
oauthlib==3.2.2
openai==1.75.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
pyobjc==11.0
pyobjc-core==11.0
pyobjc-framework-Accessibility==11.0
pyobjc-framework-Accounts==11.0
pyobjc-framework-AddressBook==11.0
pyobjc-framework-AdServices==11.0
pyobjc-framework-AdSupport==11.0
pyobjc-framework-AppleScriptKit==11.0
pyobjc-framework-AppleScriptObjC==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-AppTrackingTransparency==11.0
pyobjc-framework-AudioVideoBridging==11.0
pyobjc-framework-AuthenticationServices==11.0
pyobjc-framework-AutomaticAssessmentConfiguration==11.0
pyobjc-framework-Automator==11.0
pyobjc-framework-AVFoundation==11.0
pyobjc-framework-AVKit==11.0
pyobjc-framework-AVRouting==11.0
pyobjc-framework-BackgroundAssets==11.0
pyobjc-framework-BrowserEngineKit==11.0
pyobjc-framework-BusinessChat==11.0
pyobjc-framework-CalendarStore==11.0
pyobjc-framework-CallKit==11.0
pyobjc-framework-Carbon==11.0
pyobjc-framework-CFNetwork==11.0
pyobjc-framework-Cinematic==11.0
pyobjc-framework-ClassKit==11.0
pyobjc-framework-CloudKit==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Collaboration==11.0
pyobjc-framework-ColorSync==11.0
pyobjc-framework-Contacts==11.0
pyobjc-framework-ContactsUI==11.0
pyobjc-framework-CoreAudio==11.0
pyobjc-framework-CoreAudioKit==11.0
pyobjc-framework-CoreBluetooth==11.0
pyobjc-framework-CoreData==11.0
pyobjc-framework-CoreHaptics==11.0
pyobjc-framework-CoreLocation==11.0
pyobjc-framework-CoreMedia==11.0
pyobjc-framework-CoreMediaIO==11.0
pyobjc-framework-CoreMIDI==11.0
pyobjc-framework-CoreML==11.0
pyobjc-framework-CoreMotion==11.0
pyobjc-framework-CoreServices==11.0
pyobjc-framework-CoreSpotlight==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-CoreWLAN==11.0
pyobjc-framework-CryptoTokenKit==11.0
pyobjc-framework-DataDetection==11.0
pyobjc-framework-DeviceCheck==11.0
pyobjc-framework-DeviceDiscoveryExtension==11.0
pyobjc-framework-DictionaryServices==11.0
pyobjc-framework-DiscRecording==11.0
pyobjc-framework-DiscRecordingUI==11.0
pyobjc-framework-DiskArbitration==11.0
pyobjc-framework-DVDPlayback==11.0
pyobjc-framework-EventKit==11.0
pyobjc-framework-ExceptionHandling==11.0
pyobjc-framework-ExecutionPolicy==11.0
pyobjc-framework-ExtensionKit==11.0
pyobjc-framework-ExternalAccessory==11.0
pyobjc-framework-FileProvider==11.0
pyobjc-framework-FileProviderUI==11.0
pyobjc-framework-FinderSync==11.0
pyobjc-framework-FSEvents==11.0
pyobjc-framework-GameCenter==11.0
pyobjc-framework-GameController==11.0
pyobjc-framework-GameKit==11.0
pyobjc-framework-GameplayKit==11.0
pyobjc-framework-HealthKit==11.0
pyobjc-framework-ImageCaptureCore==11.0
pyobjc-framework-InputMethodKit==11.0
pyobjc-framework-InstallerPlugins==11.0
pyobjc-framework-InstantMessage==11.0
pyobjc-framework-Intents==11.0
pyobjc-framework-IntentsUI==11.0
pyobjc-framework-IOBluetooth==11.0
pyobjc-framework-IOBluetoothUI==11.0
pyobjc-framework-IOSurface==11.0
pyobjc-framework-iTunesLibrary==11.0
pyobjc-framework-KernelManagement==11.0
pyobjc-framework-LatentSemanticMapping==11.0
pyobjc-framework-LaunchServices==11.0
pyobjc-framework-libdispatch==11.0
pyobjc-framework-libxpc==11.0
pyobjc-framework-LinkPresentation==11.0
pyobjc-framework-LocalAuthentication==11.0
pyobjc-framework-LocalAuthenticationEmbeddedUI==11.0
pyobjc-framework-MailKit==11.0
pyobjc-framework-MapKit==11.0
pyobjc-framework-MediaAccessibility==11.0
pyobjc-framework-MediaExtension==11.0
pyobjc-framework-MediaLibrary==11.0
pyobjc-framework-MediaPlayer==11.0
pyobjc-framework-MediaToolbox==11.0
pyobjc-framework-Metal==11.0
pyobjc-framework-MetalFX==11.0
pyobjc-framework-MetalKit==11.0
pyobjc-framework-MetalPerformanceShaders==11.0
pyobjc-framework-MetalPerformanceShadersGraph==11.0
pyobjc-framework-MetricKit==11.0
pyobjc-framework-MLCompute==11.0
pyobjc-framework-ModelIO==11.0
pyobjc-framework-MultipeerConnectivity==11.0
pyobjc-framework-NaturalLanguage==11.0
pyobjc-framework-NetFS==11.0
pyobjc-framework-Network==11.0
pyobjc-framework-NetworkExtension==11.0
pyobjc-framework-NotificationCenter==11.0
pyobjc-framework-OpenDirectory==11.0
pyobjc-framework-OSAKit==11.0
pyobjc-framework-OSLog==11.0
pyobjc-framework-PassKit==11.0
pyobjc-framework-PencilKit==11.0
pyobjc-framework-PHASE==11.0
pyobjc-framework-Photos==11.0
pyobjc-framework-PhotosUI==11.0
pyobjc-framework-PreferencePanes==11.0
pyobjc-framework-PushKit==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-QuickLookThumbnailing==11.0
pyobjc-framework-ReplayKit==11.0
pyobjc-framework-SafariServices==11.0
pyobjc-framework-SafetyKit==11.0
pyobjc-framework-SceneKit==11.0
pyobjc-framework-ScreenCaptureKit==11.0
pyobjc-framework-ScreenSaver==11.0
pyobjc-framework-ScreenTime==11.0
pyobjc-framework-ScriptingBridge==11.0
pyobjc-framework-SearchKit==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-SecurityFoundation==11.0
pyobjc-framework-SecurityInterface==11.0
pyobjc-framework-SensitiveContentAnalysis==11.0
pyobjc-framework-ServiceManagement==11.0
pyobjc-framework-SharedWithYou==11.0
pyobjc-framework-SharedWithYouCore==11.0
pyobjc-framework-ShazamKit==11.0
pyobjc-framework-Social==11.0
pyobjc-framework-SoundAnalysis==11.0
pyobjc-framework-Speech==11.0
pyobjc-framework-SpriteKit==11.0
pyobjc-framework-StoreKit==11.0
pyobjc-framework-Symbols==11.0
pyobjc-framework-SyncServices==11.0
pyobjc-framework-SystemConfiguration==11.0
pyobjc-framework-SystemExtensions==11.0
pyobjc-framework-ThreadNetwork==11.0
pyobjc-framework-UniformTypeIdentifiers==11.0
pyobjc-framework-UserNotifications==11.0
pyobjc-framework-UserNotificationsUI==11.0
pyobjc-framework-VideoSubscriberAccount==11.0
pyobjc-framework-VideoToolbox==11.0
pyobjc-framework-Virtualization==11.0
pyobjc-framework-Vision==11.0
pyobjc-framework-WebKit==11.0
pyparsing==3.2.3
python-dotenv==1.1.0
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
setuptools==79.0.0
slack_sdk==3.35.0
sniffio==1.3.1
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
